data-collector/app/requirements.txt
requests==2.31.0
python-dotenv
flask